bug fix, criacao da classe ValidityPolicy para controlar a vigencia
