+++ Platform.App/node_template/model/changeTrackPolicy.js
@@ -1,37 +1,18 @@
-
-var domain = require("./domain");
 var ArrayUtils = require("../utils/array")
-
-var SequelizeModelConverter = require("./sequelizeModelConverter.js");
+var ValidityPolicy = require("./validityPolicy");
 
 class ChangeTrackPolicy {
     //recebe a lista de todas as entidades de dominio que chegaram na API
     constructor(domainEntities){
-        this.entities = domainEntities;        
-    }
-
-    tracked(){
-        return this.domainOperations;
-    }
-
-    deleted(){
-        return this.domainOperations.toDelete;
-    }
-
-    updated(){
-        return this.domainOperations.toUpdate;
-    }
-
-    created(){
-        return this.domainOperations.toCreate;
-    }
-
+        this.entities = domainEntities;
+                
+    }    
+    
     apply(callback){        
         this.cascadePersist(this.entities, callback);
     }
 
     cascadePersist(entities, callback){
-        var converter = new SequelizeModelConverter();
         var arrayUtils = new ArrayUtils();        
         arrayUtils.asyncEach(entities,(item,next)=>{            
             var type = item._metadata.type;
@@ -53,30 +34,13 @@ class ChangeTrackPolicy {
         },callback);        
     }
 
-    persist(item,callback){
-        var operation = item._metadata.changeTrack;
-        if (!operation){
+    persist(item,callback){        
+        if (!item._metadata.changeTrack){
             callback(item);
         }else{
-            var type = item._metadata.type;
-            var toExecute;
-            if ("create" === operation){
-                toExecute = domain[type].create(item);
-            }else if ("update" === operation){
-                toExecute = domain[type].update(item,{where:{rid:item.rid}});
-            }else if ("destroy" === operation){
-                var clone = JSON.parse(JSON.stringify(item));
-                clone.where = {};
-                clone.where.rid = item.rid;
-                toExecute = domain[type].destroy(clone);
-            }else{
-                throw "invalid change track operation";
-            }
-            toExecute.then((result)=>{
-                callback(result.dataValues);
-            });
-        }        
-        
+            var vigencia = new ValidityPolicy();
+            vigencia.apply(item,callback);            
+        }
     }
 
 }
