--- Platform.App/node_template/model/vigencyPolicy.js
@@ -1,5 +1,0 @@
-class ValidityPolicy {
-
-}
-
-module.exports = ValidityPolicy;
